+++ src/Opserver.Web/Models/User.cs
@@ -16,9 +16,15 @@ namespace Opserver.Models
         /// </summary>
         public ImmutableHashSet<string> Roles { get; }
 
+        /// <summary>
+        /// Gets the <see cref="ClaimsPrincipal"/> representing the logged-in user.
+        /// </summary>
+        public ClaimsPrincipal Principal { get; }
+
         public User(SecurityProvider provider, ClaimsPrincipal principal, List<string> baseRoles, IEnumerable<StatusModule> modules)
         {
             Provider = provider;
+            Principal = principal;
             var identity = principal?.Identity;
             if (identity == null)
             {