+++ src/Opserver.Web/Security/EveryonesAnAdminProvider.cs
@@ -1,17 +1,24 @@
-﻿using Opserver.Models;
+﻿using System.Collections.Generic;
+using System.Security.Claims;
+using Opserver.Models;
 
 namespace Opserver.Security
 {
     /// <summary>
     /// Does this REALLY need an explanation?
     /// </summary>
-    public class EveryonesAnAdminProvider : SecurityProvider
+    public class EveryonesAnAdminProvider : SecurityProvider<SecuritySettings, UserNamePasswordToken>
     {
         public override string ProviderName => "Everyone's an Admin!";
-        public override SecurityProviderFlowType FlowType => SecurityProviderFlowType.None;
+        public override SecurityProviderFlowType FlowType => SecurityProviderFlowType.Username;
         public EveryonesAnAdminProvider(SecuritySettings settings) : base(settings) { }
 
-        public override bool InGroups(User user, string[] groupNames) => true;
-        public override bool ValidateToken(ISecurityProviderToken token) => true;
+        protected override bool InGroupsCore(User user, string[] groupNames) => true;
+
+        protected override bool TryValidateToken(UserNamePasswordToken token, out ClaimsPrincipal claimsPrincipal)
+        {
+            claimsPrincipal = CreateNamedPrincipal(token.UserName);
+            return true;
+        }
     }
 }