+++ src/Opserver.Web/Security/ISecurityProviderToken.cs
@@ -6,16 +6,4 @@
     public interface ISecurityProviderToken
     {
     }
-
-    public class UserNamePasswordToken : ISecurityProviderToken
-    {
-        public UserNamePasswordToken(string userName, string password)
-        {
-            UserName = userName;
-            Password = password;
-        }
-
-        public string UserName { get; }
-        public string Password { get; }
-    }
 }